Added Download Path to Config Window
The download path is now configurable via the Config window including a Browse button and validation.
The dialog to select a download path will be shown only if a path was not set already or the existing path is invalid.
Also, I fixed the Config window not loading correctly after closing it once.